requirements.txt
paho-mqtt==2.1.0
psycopg[binary]==3.2.1
fastapi==0.115.5
uvicorn==0.32.0
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2
